contacts: remove cleverness from phone number import
Don't have to be too clever with phone number import because
one can assume that it is the users' responsibility to enter a
usable phone number into their address book. Second guessing the user
may lead to wrong imports. This could be noted on the site as well,
though could be considered common sense.

As for breakage, e.g. previously the "+" signs were removed, making
international numbers unusable. SMS sending is also helped by Android,
which understands phone numbers in many formats, and e.g. the "-"
characters actually don't have to be ignored here. Spaces are removed,
however, just in case, and for cleaner presentation one the site.
.